Moved AStarShortestPathTest and AStarAdmissibleHeuristic to different packages and added standard JgraphT headers
